// src/model/pokemon.ts
export default interface Pokemon {
    pokemonNumber: number;
    attack: number;
    defense: number;
    devolution: string;
    evolution: string;
    evolutionFamily: string[];
    healthPoints: number;
    mainImage: string;
    pokemonName: string;
    pokemonType: string[];
    speed: number;
  }
  